firebase 호스팅 에러
- index, login 페이지 css와 이미지가 적용안되는 경우 발생.
- 각페이지 layout popup 컴포넌트 따로 설정.
- index code  내용 home 컴포넌트로 분리.
